requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
Brotli==1.1.0
cachetools==5.5.2
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.7
decorator==5.1.1
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
dotenv==0.9.9
executing==2.1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-vision==3.10.1
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
mutagen==1.47.0
mysqlclient==2.2.7
nest-asyncio==1.6.0
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.1
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.26.1
protobuf==5.29.4
psutil==6.1.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodomex==3.21.0
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.18.0
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pywin32==308
pyzmq==26.2.0
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
websockets==13.1
whitenoise==6.9.0
yt-dlp==2024.10.22
